Mas para el controlador ofertas. Añadido parte para crear una oferta nueva y hecho el limpiar formulario. Tambien he cambiado los OID para que sean Long todos porque daba problemas. Añadido en el controladorPrincipal un método para recuperar al voluntario que se ha identificado.

+++ src/Modelo/C_Empresa.java
@@ -1,23 +1,23 @@
 /**
- ** NOMBRE CLASE: 
+ ** NOMBRE CLASE:
  **	  C_Empresa.java
  **
  ** DESCRIPCION:
- **       Abstracción de una Empresa 
- **       
+ **       Abstracción de una Empresa
+ **
  **
  ** DESARROLLADO POR:
  **       Francisco José Legaza Bailon (PLB)
- **	   
- **        
+ **
+ **
  ** SUPERVISADO POR:
- **       
+ **
  **
  ** HISTORIA:
- ** 	000 - Abril 23, 2012 - PLB - Creacion 
- **      
+ ** 	000 - Abril 23, 2012 - PLB - Creacion
+ **
  ** NOTAS:
- **   
+ **
  **
  */
 package Modelo;
@@ -28,42 +28,42 @@ package Modelo;
  * @author PLB
  */
 public class C_Empresa extends Colaborador{
-    
-    private int OID;
+
+    private Long OID;
     private String CIF;
     private String Nombre;
     private String DireccionWeb;
     private String FAX;
-    
+
     // para facilitar la recuperacion de los campos de la interfaz en los controladores
     public static int CIF_ID = 7;
     public static int NOMBRE_ID = 8;
     public static int DIRECCION_WEB_ID = 9;
     public static int FAX_ID = 10;
-    
-    
+
+
     /**
      * Constructor por defecto
      */
     public C_Empresa() {
     }
-    
+
     /**
      * Obtener el OID de la Empresa
      * @return OID
      */
-    public int getOIDEmpresa() {
+    public Long getOIDEmpresa() {
         return OID;
     }
-    
+
     /**
      * Asignar un OID a una Empresa
      * @param OID
      */
-    public void setOIDEmpresa(int OID) {
+    public void setOIDEmpresa(Long OID) {
         this.OID = OID;
     }
-    
+
     /**
      * Obtener el CIF de la Empresa
      * @return CIF
@@ -71,7 +71,7 @@ public class C_Empresa extends Colaborador{
     public String getCIF(){
         return CIF;
     }
-    
+
     /**
      * Asignar el CIF a una Empresa
      * @param CIF
@@ -79,7 +79,7 @@ public class C_Empresa extends Colaborador{
     public void setCIF(String CIF){
         this.CIF = CIF;
     }
-    
+
     /**
      * Obtener el Nombre de la Empresa
      * @return Nombre
@@ -87,7 +87,7 @@ public class C_Empresa extends Colaborador{
     public String getNombre(){
         return Nombre;
     }
-    
+
     /**
      * Asignar el Nombre a una Empresa
      * @param Nombre
@@ -95,7 +95,7 @@ public class C_Empresa extends Colaborador{
     public void setNombre(String Nombre){
         this.Nombre = Nombre;
     }
-    
+
     /**
      * Obtener la Dirección Web de la Empresa
      * @return DirecciónWeb
@@ -103,7 +103,7 @@ public class C_Empresa extends Colaborador{
     public String getDireccionWeb(){
         return DireccionWeb;
     }
-    
+
     /**
      * Asignar la Dirección Web a una Empresa
      * @param DireccionWeb
@@ -111,7 +111,7 @@ public class C_Empresa extends Colaborador{
     public void setDireccionWeb(String DireccionWeb){
         this.DireccionWeb = DireccionWeb;
     }
-    
+
     /**
      * Obtener el Fax de la Empresa
      * @return FAX
@@ -119,7 +119,7 @@ public class C_Empresa extends Colaborador{
     public String getFax(){
         return FAX;
     }
-    
+
     /**
      * Asignar el Fax a una Empresa
      * @param Fax
@@ -127,5 +127,5 @@ public class C_Empresa extends Colaborador{
     public void setFax(String Fax){
         this.FAX = Fax;
     }
-    
+
 }
